requirements.txt
Flask
flask-cors
PyYAML
gunicorn
pytest
requests
pydantic
pydantic-settings
PyJWT
cryptography
fastapi
uvicorn
streamlit
SQLAlchemy
celery
redis
langchain
openai
